CustomException

+++ HR_Management.Application/Features/LeaveAllocation/Handlers/Commands/DeleteLeaveAllocationRequestHandler.cs
@@ -1,4 +1,6 @@
-﻿namespace HR_Management.Application.Features.LeaveAllocation.Handlers.Commands
+﻿using HR_Management.Application.Exceptions;
+
+namespace HR_Management.Application.Features.LeaveAllocation.Handlers.Commands
 {
     public class DeleteLeaveAllocationRequestHandler :
         IRequestHandler<DeleteLeaveAllocationRequest>
@@ -18,6 +20,10 @@
         public async Task Handle(DeleteLeaveAllocationRequest request, CancellationToken cancellationToken)
         {
             var leaveAllocation = await _leaveAllocationRepository.Get(request.Id);
+
+            if (leaveAllocation == null)
+                throw new NotFoundException(nameof(Domain.Entities.LeaveAllocation), request.Id);
+
             await _leaveAllocationRepository.Delete(leaveAllocation);
         }
     }
