+++ HR_Management.Application/Features/LeaveType/Handlers/Commands/DeleteLeaveTypeRequestHandler.cs
@@ -1,4 +1,6 @@
-﻿namespace HR_Management.Application.Features.LeaveType.Handlers.Commands
+﻿using HR_Management.Application.Exceptions;
+
+namespace HR_Management.Application.Features.LeaveType.Handlers.Commands
 {
     public class DeleteLeaveTypeRequestHandler :
         IRequestHandler<DeleteLeaveTypeRequest>
@@ -18,6 +20,10 @@
         public async Task Handle(DeleteLeaveTypeRequest request, CancellationToken cancellationToken)
         {
             var leaveType = await _leaveTypeRepository.Get(request.Id);
+
+            if (leaveType == null)
+                throw new NotFoundException(nameof(Domain.Entities.LeaveType), request.Id);
+
             await _leaveTypeRepository.Delete(leaveType);
         }
     }
